requirements.txt
httpx~=0.27.0
pandas~=2.2.2
requests~=2.31.0
solrpy~=1.0.0
tqdm~=4.66.4
beautifulsoup4~=4.12.3
python-dotenv